requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]

# YAML support for configuration
PyYAML>=6.0

# Ham radio tools
pyhamtools>=0.7.0
python-Levenshtein>=0.21.0  # For busted call detection

# Plotting and numerical computing
matplotlib>=3.5.0
numpy>=1.21.0

# Database
alembic>=1.12.0  # For future migrations

# Excel export
openpyxl>=3.1.0

# CLI and utilities
click>=8.1.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0


Levenshtein~=0.27.3